refactor(search applications): rename list files
